Move SQLExecutor classes to udf_utils
