+++ exasol_data_science_utils_python/preprocessing/column_preprocessor.py
@@ -4,7 +4,7 @@ from typing import List
 from exasol_data_science_utils_python.preprocessing.schema.column import Column
 from exasol_data_science_utils_python.preprocessing.schema.schema import Schema
 from exasol_data_science_utils_python.preprocessing.schema.table import Table
-from exasol_data_science_utils_python.preprocessing.sql_executor import SQLExecutor
+from exasol_data_science_utils_python.udf_utils.sql_executor import SQLExecutor
 
 
 class ColumnPreprocessor(ABC):
